// src/types/instrument.ts
export interface InstrumentData {
  cliente?: string;
  proprietario?: string;
  instrumento?: string;
  codigo?: string;
  fabricante?: string;
  modelo?: string;
  faixa?: string;
  resolucao?: string;
  erro_admissivel?: string;
  periodicidade?: string;
  local_utilizacao?: string;
  unidade?: string;
}